pubsub/requirements.txt
Flask==3.0.0
pytest==7.0.1; python_version > "3.0"
# pin pytest to 4.6.11 for Python2.
pytest==7.0.1; python_version < "3.0"
gunicorn==20.1.0
Werkzeug==3.0.1
google-cloud-storage==2.12.0
google-cloud-vision==3.4.5
gunicorn==21.2.0
Wand==0.6.13

